+++ src/programs/libpfasst_swe_sphere/chooks.cpp
@@ -1,5 +1,5 @@
 #include <mpi.h>
-#include "SphereDataVars.hpp"
+#include "../libpfasst_interface/SphereDataVars.hpp"
 #include "SphereDataCtx.hpp"
 #include "ceval.hpp"
